Match prediction functionality
- Ability to predict a result in each column for each team

For every column in each team, the outliers will be removed and the average will be calculated. Based on the positivity/negativity of the columns (eg. balls scored vs balls missed), a singular numeric value will be calculated for each team using the average of each column. This numeric value will serve as the ranking to predict a win or loss. The averaged column data will serve as the predicted match data.